+++ CommonplaceBookApp/DocumentTableViewCell.swift
@@ -1,5 +1,6 @@
 // Copyright © 2017-present Brian's Brain. All rights reserved.
 
+import Combine
 import SnapKit
 import UIKit
 
@@ -33,15 +34,46 @@ final class DocumentTableViewCell: UITableViewCell {
 
     backgroundColor = UIColor.systemBackground
     divider.backgroundColor = UIColor.separator
+
+    timestampUpdatingPipeline = Just(Date())
+      .merge(with: Timer.publish(every: .minute, on: .main, in: .common).autoconnect())
+      .combineLatest(documentModifiedTimestampSubject)
+      .sink { [weak self] currentTime, modifiedTime in
+        self?.updateAgeLabel(currentTime: currentTime, modifiedTime: modifiedTime)
+      }
   }
 
   required init?(coder aDecoder: NSCoder) {
     fatalError("init(coder:) has not been implemented")
   }
 
+  var documentModifiedTimestamp: Date? {
+    get { documentModifiedTimestampSubject.value }
+    set { documentModifiedTimestampSubject.value = newValue }
+  }
+
   let titleLabel = UILabel(frame: .zero)
   let detailLabel = UILabel(frame: .zero)
-  let ageLabel = UILabel(frame: .zero)
+
+  private var timestampUpdatingPipeline: AnyCancellable?
+  private let ageLabel = UILabel(frame: .zero)
 
   private let divider = UIView(frame: .zero)
+  private var documentModifiedTimestampSubject = CurrentValueSubject<Date?, Never>(nil)
+
+  private func updateAgeLabel(currentTime: Date, modifiedTime: Date?) {
+    guard let modifiedTime = modifiedTime else {
+      ageLabel.text = ""
+      return
+    }
+    let dateDelta = currentTime.timeIntervalSince(modifiedTime)
+    ageLabel.attributedText = NSAttributedString(
+      string: DateComponentsFormatter.age.string(from: dateDelta) ?? "",
+      attributes: [
+        .font: UIFont.preferredFont(forTextStyle: .caption1),
+        .foregroundColor: UIColor.secondaryLabel,
+      ]
+    )
+    setNeedsLayout()
+  }
 }